requirements.txt
opencv-python
pillow
python-dotenv
flask
line-bot-sdk
gunicorn
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
